WIP fs writer (does not work)
